+++ src/components/Status/Status.tsx
@@ -47,8 +47,7 @@ import { useAssetById } from 'store/assets'
 import { Amount } from 'components/Amount/Amount'
 import { QRCode } from 'components/QRCode/QRCode'
 import { useCopyToClipboard } from 'hooks/useCopyToClipboard'
-import { bnOrZero } from 'lib/bignumber/bignumber'
-import { fromBaseUnit } from 'lib/bignumber/conversion'
+import { bnOrZero, fromBaseUnit } from 'lib/bignumber/bignumber'
 import type { SwapFormData } from 'types/form'
 
 import { CHAINFLIP_COMMISSION_BPS } from '../../lib/const'
