+++ src/components/TradeInput.tsx
@@ -29,8 +29,7 @@ import { NumericFormat } from 'react-number-format'
 import { useNavigate } from 'react-router'
 import { useAssetById } from 'store/assets'
 import { useDebounce } from 'hooks/useDebounce'
-import { bn, bnOrZero } from 'lib/bignumber/bignumber'
-import { fromBaseUnit, toBaseUnit } from 'lib/bignumber/conversion'
+import { bn, bnOrZero, fromBaseUnit, toBaseUnit } from 'lib/bignumber/bignumber'
 import { mixpanel, MixPanelEvent } from 'lib/mixpanel'
 import { validateAddress } from 'lib/validation'
 import type { Asset } from 'types/Asset'
